break out clock source types
