+++ apps/server/src/services/Clock.ts
@@ -1,25 +1,19 @@
-import easymidi from 'easymidi';
+import { MidiClock } from './CLock.midi.js';
+import { SystemClock } from './Clock.system.js';
+
 enum Source {
   System = 'system',
   MIDI = 'MIDI',
 }
 
+export interface ClockInterface {
+  getTime: () => number;
+}
+
 class Clock {
   private static instance: Clock;
-  private readonly source: Source;
-  private readonly midiIn;
-  private midiState = {
-    rate: 0,
-    hourMsbit: 0,
-    hourLsbits: 0,
-    minuteMsbits: 0,
-    minuteLsbits: 0,
-    secondMsbits: 0,
-    secondLsbits: 0,
-    frameMsbit: 0,
-    frameLsbits: 0,
-    total: 0,
-  };
+  private readonly source: ClockInterface;
+  private readonly sytemTime: ClockInterface;
 
   constructor(source?: Source) {
     if (Clock.instance) {
@@ -27,52 +21,15 @@ class Clock {
     }
 
     Clock.instance = this;
-
-    this.source = source || Source.System;
-    if (source === Source.MIDI) {
-      this.midiIn = new easymidi.Input(easymidi.getInputs()[0]);
-      this.midiIn.on('mtc', (mtc) => {
-        switch (mtc.type) {
-          case 0: {
-            this.midiState.frameLsbits = mtc.value & 0b00001111;
-            this.midiState.total =
-              ((this.midiState.frameMsbit + this.midiState.frameLsbits) / this.midiState.rate) * 1000 +
-              (this.midiState.secondMsbits + this.midiState.secondLsbits) * 1000 +
-              (this.midiState.minuteMsbits + this.midiState.minuteLsbits) * 60000 +
-              (this.midiState.hourMsbit + this.midiState.hourLsbits) * 3600000;
-            break;
-          }
-          case 1: {
-            this.midiState.frameMsbit = (mtc.value & 0b00000001) << 4;
-            break;
-          }
-          case 2: {
-            this.midiState.secondLsbits = mtc.value & 0b00001111;
-            break;
-          }
-          case 3: {
-            this.midiState.secondMsbits = (mtc.value & 0b00000011) << 4;
-            break;
-          }
-          case 4: {
-            this.midiState.minuteLsbits = mtc.value & 0b00001111;
-            break;
-          }
-          case 5: {
-            this.midiState.minuteMsbits = (mtc.value & 0b00000011) << 4;
-            break;
-          }
-          case 6: {
-            this.midiState.hourLsbits = mtc.value & 0b00001111;
-            break;
-          }
-          case 7: {
-            this.midiState.hourMsbit = (mtc.value & 0b00000001) << 4;
-            this.midiState.rate = [24, 25, 29.97, 30][(mtc.value >> 1) & 0b00000011];
-            break;
-          }
-        }
-      });
+    this.sytemTime = new SystemClock();
+    switch (source) {
+      case Source.System: {
+        this.source = this.sytemTime;
+        break;
+      }
+      case Source.MIDI: {
+        this.source = new MidiClock(0);
+      }
     }
   }
 
@@ -80,38 +37,11 @@ class Clock {
    * Get current time from source
    */
   timeNow(): number {
-    switch (this.source) {
-      case Source.System:
-        return this.getSystemTime();
-      case Source.MIDI:
-        return this.getMidiTime();
-      default:
-        throw new Error('Invalid time source');
-    }
+    return this.source.getTime();
   }
 
-  /**
-   * Get current time from system
-   */
-  getSystemTime() {
-    const now = new Date();
-
-    // extract milliseconds since midnight
-    let elapsed = now.getHours() * 3600000;
-    elapsed += now.getMinutes() * 60000;
-    elapsed += now.getSeconds() * 1000;
-    elapsed += now.getMilliseconds();
-    return elapsed;
-  }
-
-  /**
-   * Get current time from MIDI
-   */
-  getMidiTime() {
-    return this.midiState.total;
-    // return this.getSystemTime();
-
-    // throw new Error('Not implemented');
+  getSystemTime(): number {
+    return this.sytemTime.getTime();
   }
 }
